production_requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.1.0
qrcode[pil]==7.4.2
Pillow==10.0.1
mailtrap
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.0.0
requests==2.31.0
PyMySQL==1.1.0
cryptography==41.0.7 